profile component, added index.js
