--- client/src/pages/Dashboard/Dashboard.js
@@ -1,3 +1,0 @@
-import Dashboard from './Dashboard/Dashboard';
-
-export default Dashboard;
