--- client/src/pages/Dashboard/Profile/Profile.js
@@ -1,167 +1,0 @@
-import React from 'react';
-import UserAvatar from '../../../assets/img/sammy.jpeg';
-
-const Profile = () => (
-    <div className="container">
-    <div className="row">
-        <div className="col-12">
-            <div className="media flex-wrap mb-0 align-items-center">
-                <img alt="" src={UserAvatar} className="avatar avatar-lg mb-3 mb-md-0" />
-                <div className="media-body">
-                    <form>
-                        <label className="custom-file mb-2" for="file2">
-                            <input type="file" id="file2" className="custom-file-input height-0"/>
-                            <span className="btn btn-primary"><i className="icon-upload-to-cloud">&nbsp;</i>Upload</span>
-                        </label>
-                        <div>
-                            <small>For best results, use an image at least 256px by 256px in either .jpg or .png format</small>
-                        </div>
-                    </form>
-                </div>
-            </div>
-        </div>
-        {/* <!--end of col--> */}
-    </div>
-    {/* <!--end of row--> */}
-    <hr/>
-    <div className="row mb-4">
-        <div className="col">
-            <h5>Account Details</h5>
-        </div>
-        {/* <!--end of col--> */}
-    </div>
-    {/* <!--end of row--> */}
-    <div className="row">
-        <div className="col-12 col-md-4 order-md-2">
-            <div className="alert alert-info text-small" role="alert">
-                <i className="icon-shield"></i>
-                <span>
-                    Your information is always kept encrypted and can never be accessed by third parties.
-                </span>
-                <a href="/">See our privacy policy</a>
-            </div>
-        </div>
-        {/* <!--end of col--> */}
-        <div className="col-12 col-md-8 order-md-1">
-            <form className="row">
-                <div className="col-6">
-                    <div className="form-group">
-                        <label for="first-name">First Name:
-                            <span className="text-red">*</span>
-                        </label>
-                        <input className="form-control form-control-lg" type="text" name="firstName" value="Sammy" id="first-name" />
-                    </div>
-                </div>
-                <div className="col-6">
-                    <div className="form-group">
-                        <label for="last-name">Last Name:</label>
-                        <input className="form-control form-control-lg" type="text" name="lastName" value="Roberts" id="last-name" />
-                    </div>
-                </div>
-                <div className="col-12">
-                    <div className="form-group">
-                        <label for="username">Username:
-                            <span className="text-red">*</span>
-                        </label>
-                        <input className="form-control form-control-lg" type="text" name="userName" id="username" />
-                        <small>This will be displayed on your public profile</small>
-                    </div>
-                </div>
-                <div className="col-12">
-                    <div className="form-group">
-                        <label for="emailaddress">Email Address:
-                            <span className="text-red">*</span>
-                        </label>
-                        <input className="form-control form-control-lg" type="email" name="emailAddress" id="emailaddress" />
-                        <small>Used to log in to your account</small>
-                    </div>
-                </div>
-                <div className="col-12">
-                    <div className="form-group">
-                        <button className="btn btn-secondary" type="submit" disabled>Save changes</button>
-                    </div>
-                </div>
-            </form>
-        </div>
-        {/* <!--end of col--> */}
-    </div>
-    {/* <!--end of row--> */}
-    <hr/>
-    <div className="row mb-4">
-        <div className="col">
-            <h5>Profile Information</h5>
-        </div>
-        {/* <!--end of col--> */}
-    </div>
-    {/* <!--end of row--> */}
-    <div className="row">
-        <div className="col-12 col-md-4 order-md-2">
-            <div className="alert alert-info text-small" role="alert">
-                <i className="icon-user"></i>
-                <span>
-                    This information will appear on your public profile. A detailed public profile helps users with similar professional interests and location to connect with you.
-                </span>
-                <a href="/">View your public profile</a>
-            </div>
-        </div>
-        {/* <!--end of col--> */}
-        <div className="col-12 col-md-8 order-md-1">
-            <form className="row">
-                <div className="col-6">
-                    <div className="form-group">
-                        <label for="city">City:</label>
-                        <input className="form-control form-control-lg" type="text" name="profileCity" value="Irvine" id="city" />
-                    </div>
-                </div>
-                <div className="col-6">
-                    <div className="form-group">
-                        <label for="state">State:</label>
-                        <input className="form-control form-control-lg" type="text" name="profileState" value="CA" id="state" />
-                    </div>
-                </div>
-                <div className="col-12">
-                    <div className="form-group">
-                        <label for="website">Website:</label>
-                        <input className="form-control form-control-lg" type="text" name="profileWebsite" id="website" />
-                        <small>This will be displayed on your public profile</small>
-                    </div>
-                </div>
-                <div className="col-12">
-                    <div className="form-group">
-                        <label for="twitter">Twitter Handle:</label>
-                        <div className="input-group input-group-lg">
-                            <div className="input-group-prepend">
-                                <span className="input-group-text">@</span>
-                            </div>
-                            <input type="text" className="form-control" placeholder="Username" aria-label="twitterUsername" name="profileTwitter" id="twitter"/>
-                        </div>
-                    </div>
-                </div>
-                <div className="col-12">
-                    <div className="form-group">
-                        <label for="about-me">About me:</label>
-                        <textarea className="form-control form-control-lg" name="profileBio" rows="4" id="about-me"></textarea>
-                    </div>
-                </div>
-                <div className="col-12">
-                    <div className="form-group">
-                        <div className="custom-control custom-checkbox custom-checkbox-switch">
-                            <input type="checkbox" className="custom-control-input" name="profilePublic" id="public" />
-                            <label className="custom-control-label" for="public">Make my profile public</label>
-                        </div>
-                    </div>
-                </div>
-                <div className="col-12">
-                    <div className="form-group">
-                        <button className="btn btn-secondary" type="submit" disabled>Save changes</button>
-                    </div>
-                </div>
-            </form>
-        </div>
-        {/* <!--end of col--> */}
-    </div>
-    {/* <!--end of row--> */}
-</div>
-)
-
-export default Profile;
